# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_fluentfox_api_pkg_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorBody'
    type: object
  internal_admin.AdminUserResponse:
    properties:
      banned:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  internal_admin.BanUserRequest:
    properties:
      reason:
        maxLength: 500
        type: string
    required:
    - reason
    type: object
  internal_admin.StatsResponse:
    properties:
      active_today:
        type: integer
      chapters_published:
        type: integer
      total_reviews:
        type: integer
      total_users:
        type: integer
    type: object
  internal_auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    Japanese language learning API — chapters, SRS flashcard reviews, quizzes, XP levelling, streaks, and in-app shop.
    All successful responses are wrapped in {"data": ...}. Paginated responses also include {"meta": {...}}.
    All error responses follow {"error": {"code": "...", "message": "..."}}.
  title: FluentFox API
  version: "1.0"
paths:
  /admin/stats:
    get:
      description: Returns platform-wide statistics (total users, active today, reviews,
        published chapters). Requires admin role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_admin.StatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get platform stats
      tags:
      - admin
  /admin/users:
    get:
      description: Returns a paginated list of all users with admin-only fields. Requires
        admin role.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Results per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_admin.AdminUserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - admin
  /admin/users/{id}/ban:
    post:
      consumes:
      - application/json
      description: Bans a user account with a mandatory reason. Requires admin role.
      parameters:
      - description: Target user UUID
        in: path
        name: id
        required: true
        type: string
      - description: Ban reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_admin.BanUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ban a user
      tags:
      - admin
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset link to the email if an account exists.
        Always returns 200 to prevent email enumeration.
      parameters:
      - description: Account email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates credentials and returns access + refresh tokens.
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the supplied refresh token.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log out
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new access + refresh token
        pair.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a user account and sends an email verification link.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Register a new account
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password using a one-time token from the reset
        email.
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Token invalid, expired, or already used
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Reset password
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Marks the user's email as verified using the token sent during
        registration.
      parameters:
      - description: Verification token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Token invalid, expired, or already used
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      summary: Verify email address
      tags:
      - auth
  /chapters:
    get:
      description: Returns a paginated list of published chapters, optionally filtered
        by JLPT level.
      parameters:
      - description: JLPT level filter
        enum:
        - N5
        - N4
        - N3
        - N2
        - N1
        in: query
        name: level
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Results per page (default: 20)'
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_chapter.ChapterResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List chapters
      tags:
      - chapters
  /chapters/{slug}:
    get:
      description: Returns a single chapter with its grammar concepts and cultural
        insights.
      parameters:
      - description: Chapter slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chapter.ChapterDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get chapter detail
      tags:
      - chapters
  /progress:
    get:
      description: Returns a summary of the user's progress across all chapters and
        vocabulary.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_progress.OverallProgressResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get overall progress
      tags:
      - progress
  /quiz/sessions:
    post:
      consumes:
      - application/json
      description: Creates a new quiz session of the given type. Optionally scoped
        to a specific chapter.
      parameters:
      - description: Quiz type, optional chapter, card count
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_quiz.StartSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_quiz.SessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start a quiz session
      tags:
      - quiz
  /quiz/sessions/{id}/answers:
    post:
      consumes:
      - application/json
      description: Submits an answer for one item in an active quiz session. Returns
        correctness and XP earned.
      parameters:
      - description: Quiz session UUID
        in: path
        name: id
        required: true
        type: string
      - description: Answer payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_quiz.SubmitAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_quiz.AnswerResultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit a quiz answer
      tags:
      - quiz
  /quiz/sessions/{id}/finish:
    post:
      description: Marks the quiz session as completed and returns a summary with
        accuracy and total XP earned.
      parameters:
      - description: Quiz session UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_quiz.SessionSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Finish a quiz session
      tags:
      - quiz
  /shop:
    get:
      description: Returns all currently available items in the XP shop.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_shop.ShopItemResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List shop items
      tags:
      - shop
  /shop/inventory:
    get:
      description: Returns all items the authenticated user has purchased.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_shop.InventoryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user inventory
      tags:
      - shop
  /shop/purchase:
    post:
      consumes:
      - application/json
      description: Spends XP to purchase an item. Returns inventory ID of the new
        item.
      parameters:
      - description: Item ID to purchase
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_shop.PurchaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_shop.PurchaseResponse'
        "400":
          description: Insufficient XP or item unavailable
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Purchase a shop item
      tags:
      - shop
  /srs/due:
    get:
      description: Returns cards that are due for review today, ordered by due date.
        Limit defaults to the user's deck setting.
      parameters:
      - description: Maximum number of cards to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_srs.CardResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get due SRS cards
      tags:
      - srs
  /srs/due/count:
    get:
      description: Returns counts of new cards and review cards due today.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_srs.DueCountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get due card count
      tags:
      - srs
  /srs/review:
    post:
      consumes:
      - application/json
      description: Records a review rating (1=Again, 2=Hard, 3=Good, 4=Easy) and reschedules
        the card using the FSRS algorithm. Awards XP.
      parameters:
      - description: Card ID and rating
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_srs.SubmitReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_srs.ReviewResultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit a card review
      tags:
      - srs
  /streak:
    get:
      description: Returns the user's current and longest streak, freeze count, and
        whether the streak is alive today.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_streak.StreakResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get streak
      tags:
      - streak
  /users/me:
    get:
      description: Returns the authenticated user's account details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
  /users/me/change-password:
    post:
      consumes:
      - application/json
      description: Changes the authenticated user's password after verifying the current
        one.
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Current password incorrect
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_fluentfox_api_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your_access_token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"